# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Farmers Market API",
    "description": "Backend API for the Farm-Connect Farmers Market platform (Nigeria)",
    "version": "2.1.0"
  },
  "paths": {
...
    "/wallet/balance": {
      "get": {
        "tags": [
          "wallet"
        ],
        "summary": "Get Balance",
        "description": "Return the current wallet balance for the authenticated user.",
        "operationId": "get_balance_wallet_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wallet/topup": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Topup Wallet",
        "description": "Credit the customer's wallet by the specified amount.\nIn production, validate the payment gateway reference before crediting.",
        "operationId": "topup_wallet_wallet_topup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wallet/paystack/init": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Init Paystack Topup",
        "description": "Initialize a Paystack transaction for a wallet top-up.",
        "operationId": "init_paystack_topup_wallet_paystack_init_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaystackInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wallet/paystack/verify": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Verify Paystack Topup",
        "description": "Verify Paystack transaction and credit wallet on success.",
        "operationId": "verify_paystack_topup_wallet_paystack_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/wallet/payout": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Request Payout",
        "description": "Vendor: Request a payout of available wallet balance to their bank account.\nDeducts the requested amount from the vendor's wallet and logs the transaction.",
        "operationId": "request_payout_wallet_payout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wallet/history": {
      "get": {
        "tags": [
          "wallet"
        ],
        "summary": "Get Wallet History",
        "description": "Return the transaction history for the authenticated user's wallet.",
        "operationId": "get_wallet_history_wallet_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaystackInitRequest": {
        "properties": {
          "amount_kobo": {
            "type": "integer",
            "title": "Amount Kobo"
          }
        },
        "type": "object",
        "required": [
          "amount_kobo"
        ],
        "title": "PaystackInitRequest"
      },
...
      "TopupRequest": {
        "properties": {
          "amount_kobo": {
            "type": "integer",
            "title": "Amount Kobo"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          }
        },
        "type": "object",
        "required": [
          "amount_kobo",
          "reference"
        ],
        "title": "TopupRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}